refactor: refactor ipfs storage file names
