+++ packages/ipfs-storage/src/index.ts
@@ -1,5 +1,5 @@
-export { IpfsMetadataStorage } from "./ipfs/ipfsMetadataStorage";
-export { BaseIpfsStorage } from "./ipfs/baseIpfsStorage";
+export { IpfsMetadataStorage } from "./ipfs/metadata";
+export { BaseIpfsStorage } from "./ipfs/base";
 
 export * as constants from "./constants";
 export * as validation from "./validation";
